refactor: Move `InternalUtils` to internal package
