+++ src/test/java/jp/co/moneyforward/autotest/ut/testclasses/StateEnsuringTechniqueByFallingBackDependencies.java
@@ -14,7 +14,7 @@ import java.util.Objects;
 import java.util.function.Function;
 
 import static jp.co.moneyforward.autotest.framework.testengine.PlanningStrategy.DEPENDENCY_BASED;
-import static jp.co.moneyforward.autotest.framework.utils.InternalUtils.createContext;
+import static jp.co.moneyforward.autotest.framework.internal.InternalUtils.createContext;
 
 @AutotestExecution(
     defaultExecution = @Spec(
